requirements.txt
django_crispy_forms
crispy_bootstrap5
pillow